fix(email-first): Redirect to correct settings page after authentication
If a user tried to go to, e.g., /settings/avatar/change
without having a session, they were asked to sign in.
After sign-in, they were redirected back to /settings
rather than /settings/avatar/change because redirectTo
was not correctly propagated. Instead of propagating
the value in the view->view model, rely on the value
stored in the relier.

This also changes the default behavior of an unauthenticdated
user - instead of sending the user to `/signin`,
send the user to `/`

fixes #3270